review(547): address review findings end-to-end
Closes BLOCKERs + MAJORs from the pr-code-reviewer report and the
high-leverage NITs (migration tests, daemon tests, dead code).

BLOCKERs
- (1) cmd/gatewayd-public: loadSecretFile reads the Hetzner DNS
  token CONTENTS (not the path); previously certmagic's DNS-01
  solver was being handed the literal /etc/faas/... path and every
  ACME challenge would have failed. Mirrors cmd/gatewayd:510-512.
- (2) cmd/gatewayd-public: default control addr is 127.0.0.1:9090
  (not ':9090'); the legacy cmd/gatewayd still binds :9090 so the
  two daemons would have collided. New env knob:
  FAAS_PUBLIC_CONTROL_ADDR.

MAJORs
- (3) cmd/gatewayd-public: setupReadiness now registers a real
  pgProbeSig and mirrors pgSig.Report() into it every 2 s. The
  previous bridge goroutine read pgSig.Report() and discarded the
  values, so PG outages never flipped /readyz.
- (4) pkg/gateway/internal_proxy.go: rewrote the reverse-proxy on
  http.Transport.RoundTrip (was reading req.Body into a buffer and
  copying). Now handles request bodies natively; XFF is stripped
  inbound + re-added from r.RemoteAddr (RFC 7230 §6.1 hop-by-hop
  headers stripped); copyResponseBody uses ctx-cancellable select.
- (5) cmd/gatewayd-{public,internal}/runDrain: cancellable sleep
  via select on sigCh + 1s ticker; pgStop() BEFORE the grace sleep
  so /readyz stays 503 during the entire drain.
- (6) pkg/gateway/certsync: Leader.Push + Leader.RenewAndPush
  added; previously the per-replica FileStorage could be written
  by the leader but the followers never received the new PEMs.
  Per-replica dial surface (PushDialer) so the unix-socket path is
  testable without root.
- (7) cmd/gatewayd-internal: the placeholder handler returns 200
  OK with a TEMPLATE_OK banner (the previous 503 + empty polling
  goroutine meant the daemon could be deployed but /readyz would
  never have flipped true).
- (8) pkg/gateway/internal_proxy.go: XFF trust chain — inbound
  X-Forwarded-For is stripped; only r.RemoteAddr is forwarded.
  Honors §11 trust boundary.
- (9) cmd/gatewayd-public/secrets.go: loadSecretFile enforces a
  perm allowlist (0o400/0o440/0o600/0o640) on top of stat() —
  redox-sabotage safe per the security-deep-dive pattern.

MIGRATION
- (10) 00116: dropped dead 'IS NULL OR' clause from the CHECK and
  'WHERE node_id IS NOT NULL' from the index. node_id is NOT NULL
  so the predicate was unreachable.
- (11) 00117: pg_ratelimit_counters.tokens is bigint (was double
  precision — the 'no floats near money' invariant from CLAUDE.md
  would have failed review). Added scope-source-of-truth docstring
  pointing at pkg/api/limits.go.

NEW CODE
- (12) pkg/db/notify.go: NotifyWarmHintPublished = 'warm_hint_published'
  with payload contract docstring.
- (13) migrations/00116_warm_hint_test.go: pins table shape, CHECK
  behavior, indexes, and replay-safety for slots 116/117/118.
- (14) cmd/gatewayd-{public,internal}/main_test.go: 9 + 8 unit
  tests covering secret loader, env helpers, signal bridge smoke,
  control-addr loopback invariant, banner tripwire, placeholder
  hydration flip, runDrain signature, etc.

Tests: go test -count=1 -short passes for ./cmd/gatewayd-{public,
internal}/... ./pkg/gateway/... ./pkg/db/... ./pkg/state/...
./pkg/api/...

Refs: ADR-068, PR #547 review.

+++ migrations/00117_pg_ratelimit.sql
@@ -1,7 +1,7 @@
 -- +goose Up
 -- +goose StatementBegin
 --
--- 00117_pg_ratelimit.sql — Tier A7 edge split (ADR-070).
+-- 00117_pg_ratelimit.sql — Tier A7 edge split (ADR-068 item 7).
 --
 -- Today's per-process token bucket (pkg/gateway/ratelimit.go) is
 -- correct for one-box (one gatewayd, one bucket per app). After the
@@ -18,13 +18,15 @@
 -- hashtext((scope, subject_id)::record::text)) = 0 … RETURNING
 -- tokens` so two replicas contending on the same row serialise.
 --
--- Bench (ADR-040 follow-up): P50 0.8 ms, P99 3.2 ms on EX44.
+-- Bench (ADR-068 bench follow-up): P50 0.8 ms, P99 3.2 ms on EX44.
 --
 -- Schema invariants:
 --   - PRIMARY KEY (scope, subject_id, plan) so the ON CONFLICT
 --     clause has a deterministic target.
 --   - CHECK on scope value to keep the enum closed; future scopes
---     need a fresh ADR + migration.
+--     need a fresh ADR + migration. The canonical list lives in
+--     pkg/api/limits.go (rate-limit scope constants); this CHECK
+--     is the SQL-side duplicate for defense-in-depth.
 --   - CHECK on tokens >= 0 — the consume path uses NOT tokens >= 0
 --     to detect overflow at the SQL layer; tokens should never go
 --     negative because the limiter math is positive.
@@ -32,6 +34,10 @@
 --   - Partial index on (subject_id) WHERE scope = 'app' (the hot
 --     read path) — the per-account lookup is rarer so the partial
 --     index covers the common case.
+--   - tokens is bigint (no fractional tokens today). If a future
+--     ADR adds fractional refill (e.g. 0.5 tokens per second),
+--     rename to numeric(20,4) in a follow-up migration. The
+--     "no floats near money" invariant from CLAUDE.md applies.
 --
 -- Replay-safe (ADR-041): CREATE TABLE IF NOT EXISTS + ADD CONSTRAINT
 -- IF NOT EXISTS via DO-block (same pattern as 00116).
@@ -40,7 +46,7 @@ CREATE TABLE IF NOT EXISTS pg_ratelimit_counters (
     scope        text             NOT NULL CHECK (scope IN ('app', 'account')),
     subject_id   uuid             NOT NULL,
     plan         text             NOT NULL CHECK (plan IN ('free', 'hobby', 'pro', 'scale')),
-    tokens       double precision NOT NULL CHECK (tokens >= 0),
+    tokens       bigint           NOT NULL CHECK (tokens >= 0),
     last_refill  timestamptz      NOT NULL DEFAULT now(),
     PRIMARY KEY (scope, subject_id, plan)
 );
